Move liftAssets inside of test file
