+++ packages/contentful-to-structured-text/__tests__/all.test.ts
@@ -3,7 +3,6 @@
 
 import { richTextToStructuredText, datoToContentfulMarks } from '../src';
 import { allowedChildren, validate } from 'datocms-structured-text-utils';
-import liftAssets from './liftAssets';
 
 describe('contentful-to-structured-text', () => {
   it('works with empty rich text', async () => {
@@ -1244,7 +1243,77 @@ describe('contentful-to-structured-text', () => {
     });
 
     it('split nodes with images', async () => {
+      function liftAssets(richText) {
+        const visit = (node, cb, index = 0, parents = []) => {
+          if (node.content && node.content.length > 0) {
+            node.content.forEach((child, index) => {
+              visit(child, cb, index, [...parents, node]);
+            });
+          }
+
+          cb(node, index, parents);
+        };
+
+        const liftedImages = new WeakSet();
+
+        visit(richText, (node, index, parents) => {
+          if (
+            !node ||
+            node.nodeType !== 'embedded-asset-block' ||
+            liftedImages.has(node) ||
+            parents.length === 1 // is a top level asset
+          ) {
+            return;
+          }
+
+          const imgParent = parents[parents.length - 1];
+
+          imgParent.content.splice(index, 1);
+
+          let i = parents.length;
+          let splitChildrenIndex = index;
+          const contentAfterSplitPoint = [];
+
+          while (--i > 0) {
+            const parent = parents[i];
+            const parentsParent = parents[i - 1];
+
+            contentAfterSplitPoint = parent.content.splice(splitChildrenIndex);
+
+            splitChildrenIndex = parentsParent.content.indexOf(parent);
+
+            let nodeInserted = false;
+
+            if (i === 1) {
+              splitChildrenIndex += 1;
+              parentsParent.content.splice(splitChildrenIndex, 0, node);
+              liftedImages.add(node);
+
+              nodeInserted = true;
+            }
+
+            splitChildrenIndex += 1;
+
+            if (contentAfterSplitPoint.length > 0) {
+              parentsParent.content.splice(splitChildrenIndex, 0, {
+                ...parent,
+                content: contentAfterSplitPoint,
+              });
+            }
+            // Remove the parent if empty
+            if (parent.content.length === 0) {
+              splitChildrenIndex -= 1;
+              parentsParent.content.splice(
+                nodeInserted ? splitChildrenIndex - 1 : splitChildrenIndex,
+                1,
+              );
+            }
+          }
+        });
+      }
+
       const handlers = {
+        // eslint-disable-next-line @typescript-eslint/no-unused-vars
         'embedded-asset-block': async (createNode, node, context) => {
           const item = '123';
           return createNode('block', {
